+++ src/app/projects/states/tutor-discussion/tutor-discussion.component.ts
@@ -245,7 +245,11 @@ export class TutorDiscussionComponent implements AfterViewInit {
     const selectedTasks = this.tasksList.selectedOptions.selected;
     for (const taskOption of selectedTasks) {
       const task = taskOption.value as Task;
-      task.updateTaskStatus(status, true);
+      if (task.definition.assessInPortfolioOnly) {
+        task.updateTaskStatus(status === 'complete' ? 'working_on_it' : status, true);
+      } else {
+        task.updateTaskStatus(status, true);
+      }
     }
   }